docs: add more documentation for abstracting logic
